# backend/internal/httpserver/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Open Model Gateway API
  version: 0.1.0
  description: OpenAI-compatible gateway API with admin and user surfaces.
servers:
- url: /
paths:
...
  /v1/batches:
    get:
      tags:
      - Batches
      summary: GET /v1/batches
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
      security:
      - apiKeyAuth: []
      description: List batches. Requires API key authentication.
    post:
      tags:
      - Batches
      summary: POST /v1/batches
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
      security:
      - apiKeyAuth: []
      description: Create batches. Requires API key authentication.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchCreateRequest'
  /v1/batches/{id}:
    get:
      tags:
      - Batches
      summary: GET /v1/batches/{id}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
      security:
      - apiKeyAuth: []
      description: Get batches. Requires API key authentication.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Resource identifier.
  /v1/batches/{id}/errors:
    get:
      tags:
      - Batches
      summary: GET /v1/batches/{id}/errors
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
      security:
      - apiKeyAuth: []
      description: Download the batch error file. Requires API key authentication.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Resource identifier.
  /v1/batches/{id}/output:
    get:
      tags:
      - Batches
      summary: GET /v1/batches/{id}/output
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
      security:
      - apiKeyAuth: []
      description: Download the batch output file. Requires API key authentication.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Resource identifier.
...
  /v1/batches/{id}/cancel:
    post:
      tags:
      - Batches
      summary: POST /v1/batches/{id}/cancel
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
      security:
      - apiKeyAuth: []
      description: Cancel the batch job. Requires API key authentication.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Resource identifier.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericRequest'
...
  /admin/batches:
    get:
      tags:
      - Admin Batches
      summary: GET /admin/batches
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
      security:
      - bearerAuth: []
      description: List batches. Requires admin bearer token.
  /admin/batches/{batchID}/errors:
    get:
      tags:
      - Admin Batches
      summary: GET /admin/batches/{batchID}/errors
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
      security:
      - bearerAuth: []
      description: Download the batch error file. Requires admin bearer token.
      parameters:
      - name: batchID
        in: path
        required: true
        schema:
          type: string
        description: Batch identifier.
  /admin/batches/{batchID}/output:
    get:
      tags:
      - Admin Batches
      summary: GET /admin/batches/{batchID}/output
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
      security:
      - bearerAuth: []
      description: Download the batch output file. Requires admin bearer token.
      parameters:
      - name: batchID
        in: path
        required: true
        schema:
          type: string
        description: Batch identifier.
...
  /admin/settings/batches:
    get:
      tags:
      - Admin Settings
      summary: GET /admin/settings/batches
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
      security:
      - bearerAuth: []
      description: List batches. Requires admin bearer token.
    put:
      tags:
      - Admin Settings
      summary: PUT /admin/settings/batches
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
      security:
      - bearerAuth: []
      description: Replace batches. Requires admin bearer token.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericRequest'
...
  /admin/tenants/{tenantID}/batches:
    get:
      tags:
      - Admin Tenants
      summary: GET /admin/tenants/{tenantID}/batches
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
      security:
      - bearerAuth: []
      description: Get batches. Requires admin bearer token.
      parameters:
      - name: tenantID
        in: path
        required: true
        schema:
          type: string
        description: Tenant identifier.
  /admin/tenants/{tenantID}/batches/{batchID}:
    get:
      tags:
      - Admin Tenants
      summary: GET /admin/tenants/{tenantID}/batches/{batchID}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
      security:
      - bearerAuth: []
      description: Get batches. Requires admin bearer token.
      parameters:
      - name: tenantID
        in: path
        required: true
        schema:
          type: string
        description: Tenant identifier.
      - name: batchID
        in: path
        required: true
        schema:
          type: string
        description: Batch identifier.
  /admin/tenants/{tenantID}/batches/{batchID}/errors:
    get:
      tags:
      - Admin Tenants
      summary: GET /admin/tenants/{tenantID}/batches/{batchID}/errors
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
      security:
      - bearerAuth: []
      description: Download the batch error file. Requires admin bearer token.
      parameters:
      - name: tenantID
        in: path
        required: true
        schema:
          type: string
        description: Tenant identifier.
      - name: batchID
        in: path
        required: true
        schema:
          type: string
        description: Batch identifier.
  /admin/tenants/{tenantID}/batches/{batchID}/output:
    get:
      tags:
      - Admin Tenants
      summary: GET /admin/tenants/{tenantID}/batches/{batchID}/output
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
      security:
      - bearerAuth: []
      description: Download the batch output file. Requires admin bearer token.
      parameters:
      - name: tenantID
        in: path
        required: true
        schema:
          type: string
        description: Tenant identifier.
      - name: batchID
        in: path
        required: true
        schema:
          type: string
        description: Batch identifier.
...
  /admin/batches/{batchID}/cancel:
    post:
      tags:
      - Admin Batches
      summary: POST /admin/batches/{batchID}/cancel
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
      security:
      - bearerAuth: []
      description: Cancel the batch job. Requires admin bearer token.
      parameters:
      - name: batchID
        in: path
        required: true
        schema:
          type: string
        description: Batch identifier.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericRequest'
...
  /admin/tenants/{tenantID}/batches/{batchID}/cancel:
    post:
      tags:
      - Admin Tenants
      summary: POST /admin/tenants/{tenantID}/batches/{batchID}/cancel
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
      security:
      - bearerAuth: []
      description: Cancel the batch job. Requires admin bearer token.
      parameters:
      - name: tenantID
        in: path
        required: true
        schema:
          type: string
        description: Tenant identifier.
      - name: batchID
        in: path
        required: true
        schema:
          type: string
        description: Batch identifier.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericRequest'
...
  /user/tenants/{tenantID}/batches:
    get:
      tags:
      - User Tenants
      summary: GET /user/tenants/{tenantID}/batches
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
      security:
      - bearerAuth: []
      description: Get batches. Requires user bearer token.
      parameters:
      - name: tenantID
        in: path
        required: true
        schema:
          type: string
        description: Tenant identifier.
  /user/tenants/{tenantID}/batches/{batchID}:
    get:
      tags:
      - User Tenants
      summary: GET /user/tenants/{tenantID}/batches/{batchID}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
      security:
      - bearerAuth: []
      description: Get batches. Requires user bearer token.
      parameters:
      - name: tenantID
        in: path
        required: true
        schema:
          type: string
        description: Tenant identifier.
      - name: batchID
        in: path
        required: true
        schema:
          type: string
        description: Batch identifier.
  /user/tenants/{tenantID}/batches/{batchID}/errors:
    get:
      tags:
      - User Tenants
      summary: GET /user/tenants/{tenantID}/batches/{batchID}/errors
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
      security:
      - bearerAuth: []
      description: Download the batch error file. Requires user bearer token.
      parameters:
      - name: tenantID
        in: path
        required: true
        schema:
          type: string
        description: Tenant identifier.
      - name: batchID
        in: path
        required: true
        schema:
          type: string
        description: Batch identifier.
  /user/tenants/{tenantID}/batches/{batchID}/output:
    get:
      tags:
      - User Tenants
      summary: GET /user/tenants/{tenantID}/batches/{batchID}/output
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
      security:
      - bearerAuth: []
      description: Download the batch output file. Requires user bearer token.
      parameters:
      - name: tenantID
        in: path
        required: true
        schema:
          type: string
        description: Tenant identifier.
      - name: batchID
        in: path
        required: true
        schema:
          type: string
        description: Batch identifier.
...
  /user/tenants/{tenantID}/batches/{batchID}/cancel:
    post:
      tags:
      - User Tenants
      summary: POST /user/tenants/{tenantID}/batches/{batchID}/cancel
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
      security:
      - bearerAuth: []
      description: Cancel the batch job. Requires user bearer token.
      parameters:
      - name: tenantID
        in: path
        required: true
        schema:
          type: string
        description: Tenant identifier.
      - name: batchID
        in: path
        required: true
        schema:
          type: string
        description: Batch identifier.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericRequest'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
  schemas:
...
    GenericRequest:
      type: object
      additionalProperties: true
      example:
        note: See endpoint description or docs for required fields.
    GenericResponse:
      type: object
      additionalProperties: true
      example:
        status: ok
...
    BatchCreateRequest:
      type: object
      properties:
        input_file_id:
          type: string
        endpoint:
          type: string
        completion_window:
          type: string
        metadata:
          type: object
      example:
        input_file_id: file_123
        endpoint: /v1/chat/completions
        completion_window: 24h